accommodate more no_std builds
