+++ termwiz/Cargo.toml
@@ -58,8 +58,6 @@ docs = ["widgets", "use_serde", "image", "tmux_cc"]
 criterion.workspace = true
 env_logger.workspace = true
 k9.workspace = true
-varbincode.workspace = true
-
 
 [target."cfg(unix)".dependencies]
 signal-hook.workspace = true
